+++ lib/util/folly_utils.dart
@@ -170,7 +170,13 @@ class FollyUtils {
   ///
   ///
   ///
-  static Color? colorParse(String? text, [int? defaultColor]) {
+  static String colorHex(Color color) =>
+      color.value.toRadixString(16).toUpperCase().padLeft(8, '0');
+
+  ///
+  ///
+  ///
+  static Color? colorParse(String? text) {
     try {
       String t = text?.replaceAll('#', '').trim().toLowerCase() ?? '';
       if (!t.startsWith('0x')) {
@@ -196,7 +202,7 @@ class FollyUtils {
 
       return Color(int.parse(t));
     } on Exception catch (_) {
-      return defaultColor == null ? null : Color(defaultColor);
+      return null;
     }
   }